% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packgraph-package.R
\docType{package}
\name{packgraph-package}
\alias{packgraph}
\alias{packgraph-package}
\title{packgraph: Examine and Analyse Function Call Graphs of R Packages}
\description{
Examine and analyse function call graphs of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/packgraph}
  \item Report bugs at \url{https://github.com/mpadge/packgraph/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
