% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{pg_documentation}
\alias{pg_documentation}
\title{pg_documentation}
\usage{
pg_documentation(pkg_dir)
}
\arguments{
\item{pkg_dir}{Directory path to package (not a \code{.tar.gz} archive)}
}
\value{
A list with one item for each package function, each of which
contains a list of all of the variously delinated items defining the
function documentation.
}
\description{
Extract documentation for one package
}
