% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{pg_graph}
\alias{pg_graph}
\title{pg_graph}
\usage{
pg_graph(pkg_dir, plot = TRUE)
}
\arguments{
\item{pkg_dir}{Directory containing the package}

\item{plot}{If \code{TRUE}, plot the network using \pkg{visNetwork} which opens an
interactive browswer pane.}
}
\value{
A \code{list} of \code{nodes} and \code{edges} describing connections between the
various exported and non-exported functions of a package.
}
\description{
Construct function network graph for a package
}
