% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{pg_report}
\alias{pg_report}
\title{pg_report}
\usage{
pg_report(g)
}
\arguments{
\item{g}{A package graph objected returned from \link{pg_graph}}
}
\value{
Summary report on package structure
}
\description{
Report on package structure
}
